requirements.txt
numpy
Flask==2.0.1
keras
tensorflow-cpu
Pillow